% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanTimeSeries}
\alias{cleanTimeSeries}
\title{Clean Time Series}
\usage{
cleanTimeSeries(timeSeries_df)
}
\arguments{
\item{timeSeries_df}{- dataframe with time series to clean}
}
\description{
clean NAs from each row of a timeSeries data.frame.
if a row is too brutally NA'd to be cleaned, remove
it and set the 'omitted' attributed of the returned
timeSeries data.frame to reflect the omission.
}

