% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createCluster}
\alias{createCluster}
\title{Create Cluster}
\usage{
createCluster(noCores, logfile = "/dev/null", export = NULL, lib = NULL)
}
\arguments{
\item{noCores}{numeric - number of cores}

\item{logfile}{character - name of file to write logs to}

\item{export}{named list - items to export to each cluster}

\item{lib}{named list - libraries to export to each cluster}
}
\description{
initiate a cluster for .parallel call
}

