% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{d_dx}
\alias{d_dx}
\title{d/dx}
\usage{
d_dx(x, fx)
}
\arguments{
\item{x}{numeric vector - sequence of domain points}

\item{fx}{numeric}
}
\description{
take the derivative of a function represented as two index-aligned
vectors. returns the derivative functions represented in the same
format.
}

