% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMaxEdgesFromSource}
\alias{getMaxEdgesFromSource}
\title{Get Max Edges From Source}
\usage{
getMaxEdgesFromSource(node)
}
\arguments{
\item{node}{numeric - node number}
}
\description{
Get Maximum Edge Length From Source Node
}

