% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMinEdgesFromSource}
\alias{getMinEdgesFromSource}
\title{Get Min Edges From Source}
\usage{
getMinEdgesFromSource(node)
}
\arguments{
\item{node}{numeric - node number}
}
\description{
Get Minimum Edge Length from Source Node
}

