% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getUsageDuring}
\alias{getUsageDuring}
\title{Get Usage During (a period)}
\usage{
getUsageDuring(df, hours)
}
\arguments{
\item{df}{group segment of a data frame}

\item{hours}{- vector of hours, each sequential pair defines a period}
}
\description{
to be used with dplyr:
given a data.frame with 'Hour' as a variable
that ranges uniquely across 0-23 and 'Value' as the
quantity at each time, find the percent
of the total area between the given hour pairs 
(given as list of vector pairs)
}

