% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inquireReload}
\alias{inquireReload}
\title{Inquire Reload}
\usage{
inquireReload(yesNoQuestion, answer = NULL)
}
\arguments{
\item{yesNoQuestion}{- character yes or no question}

\item{answer}{- supply an answer to skip the process}
}
\description{
function to call the reload statement:
}

