% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mostFrequent}
\alias{mostFrequent}
\title{Most Frequent}
\usage{
mostFrequent(df, colname)
}
\arguments{
\item{df}{dataframe containing column to freq}

\item{colname}{name of the column of interest}
}
\description{
dplyr wrapper to get most frequent value of a variable
in a grouped table
}

