% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeColumn}
\alias{normalizeColumn}
\title{Normalize Column}
\usage{
normalizeColumn(df, colname)
}
\arguments{
\item{df}{- dataframe}

\item{colname}{- name of column to normalize}
}
\description{
normalize a column in a dataframe
}

