% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{performKmeans}
\alias{performKmeans}
\title{Perform K Means}
\usage{
performKmeans(df, k, .nstart)
}
\arguments{
\item{df}{- dataframe containing the data to be clustered}

\item{k}{- number of clusters}

\item{.nstart}{- number of random points from which to start clustering}
}
\description{
perform K means clustering on multidplyr/plyr-grouped data
return clustering results in the form of a data frame
}

