% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot Distribution}
\usage{
plotDistribution(df, descriptors, value_name, location)
}
\arguments{
\item{df}{- data frame containing data}

\item{descriptors}{- the grouping factors whose values will
determine the file save location}

\item{value_name}{- the name of the column containing data to be plotted}

\item{location}{- directory in which to save the plot}
}
\description{
plot and save a density distribution
}

