% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{queryAustinDB}
\alias{queryAustinDB}
\title{Query Austin Database}
\usage{
queryAustinDB(sql)
}
\arguments{
\item{sql}{chracter - sql query to be sent}
}
\description{
send a sql query to the Austin database
}

