% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smoothNAs}
\alias{smoothNAs}
\title{Smooth NAs}
\usage{
smoothNAs(timeSeries)
}
\arguments{
\item{timeSeries}{- vector timeseries to be smoothed over}
}
\description{
clear NA values from a timeseries vector by converting
them to the average of the non-na values nearest them
in the series
}

