% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sumDfColToIndex}
\alias{sumDfColToIndex}
\title{Sum Dataframe Column To Index}
\usage{
sumDfColToIndex(df, colname, new_colname)
}
\arguments{
\item{df}{- dataframe containing column to 'sum to index'}

\item{colname}{- name of the column within df to 'sum to index'}

\item{new_colname}{- name for the new 'summed' column}
}
\description{
dplyr wrapper for sumToIndex
}

