% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarizeDF}
\alias{summarizeDF}
\title{Summarize Dataframe}
\usage{
summarizeDF(df, column_name, probs = seq(0, 1, 0.25))
}
\arguments{
\item{df}{dataframe}

\item{column_name}{name of column to summarize}

\item{probs}{quantiles to summarize by}
}
\description{
quantiles, means, variances for a dataframe's variable
}

