% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sumToIndex}
\alias{sumToIndex}
\title{Sum To Index}
\usage{
sumToIndex(vector)
}
\arguments{
\item{vector}{boolean vector}
}
\description{
count number of values since the
last TRUE value in a given vector i.e.
c(FALSE, TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE)
-> c(1, 0, 0, 1, 2, 0, 1, 2, 3, 0)
note: this is biased towards 1's...
}

