% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeColumn}
\alias{normalizeColumn}
\title{Normalize Column}
\usage{
normalizeColumn(df, colname, type = "Feature Scaling")
}
\arguments{
\item{df}{- dataframe}

\item{colname}{- name of column to normalize}

\item{type}{- the type of normalization to do}
}
\description{
normalize a column in a dataframe
}

